arc-tools/grabber/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ghostgraph"
version = "0.1.0"
description = "Distributed web scraping intelligence system"
requires-python = ">=3.11"
license = "MIT"

dependencies = [
    # Web framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-multipart>=0.0.20",

    # Database
    "psycopg2-binary>=2.9.9",
    "pgvector>=0.3.0",

    # Redis
    "redis>=5.2.0",

    # Configuration
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",

    # HTTP / Stealth
    "curl_cffi>=0.7.0",
    "camoufox>=0.4.0",
    "browserforge>=1.2.0",

    # HTML parsing
    "beautifulsoup4>=4.12.0",
    "lxml>=5.3.0",
    "readability-lxml>=0.8.1",
    "selectolax>=0.3.0",
    "cssselect>=1.2.0",

    # Structured data extraction
    "extruct>=0.17.0",

    # URL handling
    "tldextract>=5.1.0",

    # Fuzzy matching
    "rapidfuzz>=3.10.0",

    # LLM output repair
    "json-repair>=0.30.0",

    # System monitoring
    "psutil>=6.1.0",

    # Process management
    "pebble>=5.0.0",

    # HTTP client (LLM)
    "httpx>=0.28.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.8.0",
    "pyright>=1.1.0",
]

[tool.hatch.build.targets.wheel]
packages = ["app", "workers", "vendor"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "SIM"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.pyright]
pythonVersion = "3.11"
typeCheckingMode = "basic"
